requirements.txt
annotated-types==0.6.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
flake8==7.0.0
Flask==3.0.2
Flask-Sijax==0.4.1
future==0.18.3
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mccabe==0.7.0
pycodestyle==2.11.1
pydantic==2.6.1
pydantic_core==2.16.2
pyflakes==3.2.0
python-decouple==3.8
requests==2.31.0
Sijax==0.3.2
six==1.16.0
typing_extensions==4.9.0
urllib3==2.2.0
Werkzeug==3.0.1
